Autoupdate of 2 submodules
*build updated to 4f82b3d39339b9f2166a2d23c9efb9f90d2a11ff
kleaf: properly handle local_defines/includes for assembly.

Previously, only if asopts is set, local_defines
and includes has an effect on assembly code. Fix
the conditional.

Bug: 392915436
Fixes: c3dba63c9637 ("kleaf: includes & local_defines written to AFLAGS_* as well.")
Change-Id: I8e984b0f9efc815ac50dafca3836fa54199ed6ee

*common-modules/virtual-device updated to f9fb5dd8c152aec830f569fcf71dbf81dc6258c6
kleaf: Fix assembly test for local_defines & includes.

Split the test to ensure that local_defines and includes
are not conditional on the existence of asopts.

Bug: 392915436
Fixes: 405f49e8e03b ("kleaf: Add DDK test for adding local_defines/includes to AFLAGS.")
Change-Id: Ic05a4a4d5dbd9538f097d0d3f4e0b36f40d61c43