ANDROID: GKI: disable CONFIG_PTP_1588_CLOCK_VMCLOCK for x86 target
In commit 205032724226 ("ptp: Add support for the AMZNC10C 'vmclock'
device"), the driver is enabled by default, yet this is not something we
want or need right now for GKI builds.  So disable it.  Later commits
make this dependent on a different config option, but keeping it
disabled is still ok for then as well.

Fixes: 205032724226 ("ptp: Add support for the AMZNC10C 'vmclock' device")
Change-Id: Id2b002f45c64a62cbb89b3afbf5e8c77f4819403
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>