ANDROID: GKI: Enable xHCI debug capability
The xHCI debug capability (DbC) is an optional functionality provided by
the xHCI host controller. When DbC is initialized and enabled, it will
present a debug device through the debug port which is fully compliant
with the USB framework and provides the equivalent of full-duplex serial
link between the debug target and a debug host.

It is required to enable ADB over USB on platforms without support for
USB device mode.

Bug: 391305431
Bug: 398809227
Change-Id: I36862716f9d8ecf1d66347763ec4f1d25dde9ff2
Signed-off-by: Albert Jakieła <jakiela@google.com>
Signed-off-by: Łukasz Bartosik <ukaszb@google.com>
Signed-off-by: Slawomir Rosek <srosek@google.com>