kleaf: set RUSTC_FORCE_RUSTC_VERSION=rustc-android
Setting this environment variable removes the dependency on the rustc
version string from symbol names. This will ensure that different kernel
branches use the same symbol names for the same symbols. It also ensures
that symbol names do not change when we respin rustc.

Rust symbol names include a disambiguator for the crate in the mangled
name, which helps distinguish different crates with the same name.
The disambiguator is a hash of the crate name and the rustc version
string, so changes to the rust version string would change the
disambiguator.

Bug: 398164112
Change-Id: I8c975d259e06f638d408c3cc39d0ca80f998085b
Signed-off-by: Alice Ryhl <aliceryhl@google.com>