ANDROID: virtio-dxgkrnl: use GFP_KERNEL
virtio-dxgkrnl is using GFP_ATOMIC in a lot of situations where it is
unnecessary. GFP_ATOMIC is only necessary in things like interrupt
handlers, where blocking on memory allocation is unacceptable. All
virtio-dxgkrnl allocations happen in work queues or in userspace ioctl
contexts where blocking on memory allocation is acceptable. This will
fix command failures that we are getting in low memory situations.

Test: ran a game
Bug: 397488281
Signed-off-by: Colin Downs-Razouk <colindr@google.com>
Change-Id: Idf0fe443ac3e65491042acc50f70ef87af944a51