kleaf: kbuild_mixed_tree_ret.outputs -> inputs.
For the caller, these are a list of inputs to be added
to the next action/step. Rename it, and change it
to a depset() to prepare for the next change.

Also reverse the conditional to prepare for the next change to reduce
duplication of code.

Also add some comments.

Bug: 402231876
Change-Id: I689effbcf15a19c52493744379936a117e2d0b97