Create the Tag enum for unions as a DERIVED location
We create internal Tag enums inside of unions for the descriminators.
We throw errors in some cases if those are handled in a backward
incompatible way.
We treat errors from INTERNAL locations as fatal errors and it causes
problems with our fuzzers.
Now we have a DERIVED location type for this case where we want the type
to be logged when there is a problem, but it is not an EXTERNAL type
from another aidl file.

Test: atest aidl_unittests
Bug: 402354469
Change-Id: I7a028b9c8040f6901750a0308238cc2ed7a830fa