GKI: Update lookup logic for symbol/section name
In recent versions of llvm-readelf the json structure has changed
and the name value comes from object['Name']['Name'] instead of
section['Name']['Value'] so try to read this if available first
and fallback to the old value if it's not present.

This commit introduces a new function to try to get the 'Name' if it's
available (recent versions) or fallback to 'Value' for older versions.

Change-Id: Ic8a809783ed67e05a27784cde0dd23b5adbea378
Signed-off-by: Ulises Mendez Martinez <umendez@google.com>