ANDROID: GKI: Enable Compute Acceleration Framework
Framework for device drivers of compute acceleration devices, such as,
but not limited to, Machine-Learning and Deep-Learning acceleration
devices. This framework is integrated with the DRM subsystem as compute
accelerators and GPUs share a lot in common and can use almost the same
infrastructure code. The CONFIG_DRM_ACCEL directly modifies DRM core
which is a GKI built-in thus it cannot be enabled from the vendor
defconfig fragment:

```
drivers/gpu/drm/drm_drv.c:#if IS_ENABLED(CONFIG_DRM_ACCEL)
```

It is required for various vendor modules: Intel iVPU, Qualcomm Cloud AI
and others.

Bug: 406034570
Change-Id: I8485b8ea76c84707b8e8198c86faa4a494c95175
Signed-off-by: Ryan Neph <ryanneph@google.com>
Signed-off-by: Slawomir Rosek <srosek@google.com>