kleaf: Output flat DLKM modules.load files to dist
When building flattened DLKM images, the distribution directory
contains the flattened DLKM image, the non-flattened DLKM image, and the
non-flattened modules.load file for the DLKM image.

This asymmetry is awkward and causes some confusion for users that want
to just use the flattened image, as the modules.load file is not
flattened.

Therefore, output the flattened modules.load files to the distribution
directory to match the artifacts for the non-flattened DLKM image, and
to avoid any confusion with regards to what the flattened modules.load
file looks like in the flattened image.

Bug: 405786400
Change-Id: Iead006d4e1271acb7a4f8671aafc87677a6a8ba0
[isaacmanjarres: resolved trivial merge conflicts related to
attribute name changes.]
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>