kleaf: Stamp ddk with our mechanism.
DDK modules are now stamped by adding .modinfo via generated code in
makefiles(), instead of via the OOT patch
commit 1bf5e5e70d85c8a47ae6495b095602efa1298d4c
("ANDROID: modules: re-introduce the MODULE_SCMVERSION config").

This is so that we can get rid of the OOT patch after kernel_module is
completely deprecated.

This patch changes how ddk_module() and ddk_submodule()'s are stamped.
The following are not affected:

- kernel_module(); it doesn't have the generated makefile, and it keeps
    follow_stamp_flag=True in _kernel_module() to maintain old behavior.
- ddk_library(): The generated makefile doesn't yield the stamp by
    setting follow_stamp_flag=False. In addition, it also sets
    follow_stamp_flag=False for _kernel_module() so Kbuild doesn't stamp
    it either.

Test: bazel test //common-modules/virtual-device:aarch64/goldfish_drivers/goldfish_pipe_test --config=stamp
Bug: 403383438
Change-Id: Ic8e98cab9868db08db4587bea274c3d47c0ae060