ANDROID: gki_defconfig: add CONFIG_CPUSETS_V1=y for x86 to match arm64
To allow devices that are still using cpusets v1 to continue working as
expected, we need to enable CONFIG_CPUSETS_V1 in GKI as of the upstream
commit 1abab1ba0775 ("cgroup/cpuset: guard cpuset-v1 code under
CONFIG_CPUSETS_V1") until we migrate devices over to v2. aosp/3284215
added that config for arm64 but missed x86, which causes failures to
mount cpuset v1 cgroup controller on Cuttlefish. Add the missing config.

Bug: 401031513
Change-Id: Id82a91c4c35324978513d371dc20e5c41694c0dc
Signed-off-by: Suren Baghdasaryan <surenb@google.com>