drivers: phy: samsung: mipi-dsim: support regmap for PMU access
Upstream supports regmap for PMU access, and it also doesn't have a
specific Exynos PMU driver exposing specific Exynos PMU APIs.

Allow this driver to use regmap in that case.

To avail of the regmap approach, the Kconfig symbol needs to be
enabled, and an appropriate DT update needs to be made. If the Kconfig
symbol isn't set, this driver behaves as before, calling the Exynos
PMU APIs directly.

Test: TH, upstream environment
Bug: 426554178
Change-Id: Id715aa614e03516f08b355b4a44638bc2943d975
Signed-off-by: André Draszik <draszik@google.com>