ANDROID: GKI: defconfig: remove assignment to CONFIG_CRC8 promptless symbol
Upstream commit [1] removed unnecessary prompt for CONFIG_CRC8, as all
the modules that need CONFIG_CRC8 already select it. With that the x86
gki_defconfig ended up assigning an invisible symbol which is ignored.
This caused some build failures as gki_defconfig did not match the output
of savedefconfig. Remove assignment to CONFIG_CRC8.

Bug: 409964359
Test: TH
Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=aa09b3223c85 [1]
Cc: Eric Biggers <ebiggers@google.com>
Change-Id: Id9baaa1b2aca4d2e405c5160d119f5c5e57a0de5
Signed-off-by: Tudor Ambarus <tudordana@google.com>