ANDROID: update KABI macros for gendwarfksyms
With the transition from genksyms to gendwarfksyms, the mechanisms for
achieving stable MODVERSIONS CRCs in the face of "safe" ABI changes are
now very different.

The macros for padding reservation / use have much the same interface,
though the variant specifically for vendors has been dropped as it was
barely used.

It is no longer possible to use compilation conditional on __GENKSYMS__
to control what is visible to symtypes generation as that now uses DWARF
information from a single compilation pass (as opposed to source from a
separate preprocessing pass).

There are specific macros to deal with common cases like member
replacement, enumerator addition and unwanted type definitions.

Note that there is as yet no support for dealing with type size changes.

Bug: 400972353
Change-Id: Ic9abadadeed6ca8f7be82165f831dab6f34e98d2
Signed-off-by: Giuliano Procida <gprocida@google.com>