ANDROID: pkvm: x86: Prevent host from donating an IOMMU mapped page
Right now, host can still donate a page even when it is DMA mapped
in the host page tables (and in the pkvm's shadow page tables). This
poses a security risk as the host device can still access the pages
which the host has already donated.

To rectify it, whenever host does a DMA mapping, pkvm should
increment the refcount (field in the struct hyp_page) of the page.
We can use this counter (refcount) as it is not being used elsewhere
for the pages owned by the host.

Now whenever the host tries to donate a page, its refcount is checked.
Donation only goes through if it's zero.

Bug: 351768668
Upstream-Task: 402758258
Test: Boot, verify by trying to donate an already DMA mapped host page.

Change-Id: I10462a78fb17be3689cac53a01f695dac6a8c4f7
Signed-off-by: Aashish Sharma <shraash@google.com>