ANDROID: KVM: arm64: Translate status codes in do_ffa_guest_features
Use ffa_to_smccc_res_prop to translate FFA_RET_SUCCESS into FFA_SUCCESS,
and so on. This mirrors the logic in do_ffa_features and ensures that
callees will interpret the results of calls to FFA_FEATURES correctly.

If the hypervisor page size is not 4K, 16K, or 64K, return not supported
for FFA_FN64_RXTX_MAP feature.

Bug: 269285339
Change-Id: I8529368a771802baed6983f3184841a322d196c6
Signed-off-by: Per Larsen <perlarsen@google.com>