FROMGIT: Bluetooth: btintel_pcie: Add additional to checks to clear TX/RX paths
Due to a hardware issue, there is a possibility that the driver may miss
an MSIx interrupt on the RX/TX data path. Since the TX and RX paths are
independent, when a TX MSIx interrupt occurs, the driver can check the
RX queue for any pending data and process it if present. The same
approach applies to the RX path.

BUG=b:400831265
TEST=Build and test bluetooth on PTLRVP

(cherry picked from commit ac1f44f12090fbe2607f704d4022b66ec8eccc2b
https://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git master)

Fixes: c2b636b3f788 ("Bluetooth: btintel_pcie: Add support for PCIe transport")
Change-Id: Ic986ac791eaf251d39ab776bb6b10823d666bb1f
Signed-off-by: Chandrashekar Devegowda <chandrashekar.devegowda@intel.com>
Signed-off-by: Kiran K <kiran.k@intel.com>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>