kleaf: properly infer Git metadata for dependent Bazel modules.
If @kleaf is not the root module, it shows up as external/kleaf~ in
the sandbox. However, it does not necessarily translate to
<workspace>/external/kleaf. This change fixes it so that we no longer
rely on the workaround that blindly converts external/kleaf~ to
<workspace>/external/kleaf when looking up Git metadata.

This fix is guarded by the --source_date_epoch_fix flag. The flag will
only be enabled on unfrozen branches. On frozen branches (android15-6.6)
users are expected to specify the flag explicitly, or they MUST checkout
@kleaf below external/kleaf.

On average, adding the --source_date_epoch_fix flag adds 0.1s of
overhead for all commands, with or without --config=stamp.
This is tested by `time bazel version`
with and without the --source_date_epoch_fix flag; the time increased
from 0.13s to 0.23s in an experiment.

Bug: 377954908
Change-Id: I15c52650f8fb0994a1d3f6e6269f37da873f100b