kleaf: Enable --source_date_epoch_fix by default.
If one encounters any issue, they may disable it with
--nosource_date_epoch_fix temporarily.

On average, adding the --source_date_epoch_fix flag adds 0.1s of
overhead for all commands, with or without --config=stamp.
This is tested by `time bazel version`
with and without the --source_date_epoch_fix flag; the time increased
from 0.13s to 0.23s in an experiment.

Bug: 377954908
Change-Id: I8c3d776557bacff2b7311d72f4a943a6c039549e