ANDROID: [TEMP] GKI: arm64: Enable mtk serial console
For the time of development we need access to the
serial console. Building the *-eng image is not enough
in case of mediatek device. This config option must be
present to get output on serial console.
It is merged only temporarily to the *desktop
branch and there is no plan to upstream it hence
bypassing the upstream divergence check.
The final solution keeps this option in a debug
vendor fragment built on demand.

Bug: 355627250
Test: tools/bazel run //icommon:kernel_aarch64_dist
Bypass-Upstream-Divergence-Tracking

Change-Id: I1f2b646bc291df64e52f3cce7dffaac894085a91
Signed-off-by: Zyta Szpak <zyta@google.com>