gbl: define EfiError and EfiResult types
Defines some Rust wrappers around EfiStatus:
 * EfiError: enum of any non-success status
 * EfiResult: Result of any status

These live in the efi_types library, which is the basic standalone EFI
lib. This intentionally does not use any other GBL libraries so that it
can be portably used in GBL as well as in UEFI firmware implementations.

This also slightly modifies the bindgen for EfiStatus so that we can
define functions on it, so corresponding libraries have been fixed up as
well. Future CLs will integrate this new code and remove some of the
duplication.

Bug: b/413021069
Change-Id: I846baa598083864543cef903f07d589ec72ebc4f