kleaf: device_module_test respects defconfig & fragments from base_kernel.
Otherwise, if the base_kernel has pre_defconfig_fragments set, the
mixed build kernel_build() does not also build these in-tree
modules with pre_defconfig_fragments, causing incompatibility between
vmlinux and the mixed build in-tree modules.

The fundamental issue is that a kernel_build() does not automatically
inherit defconfig, pre_defconfig_fragments and post_defconfig_fragments
from the base kernel. The proper fix for this is tracked in b/416803611.

Bug: 415825866
Bug: 416803611
Change-Id: Ied80915a5a1da35a8548bc53586d7d0500e5a0d8