kleaf: kernel_config adds DefconfigFragmentsListInfo.
This is a private helper provider that contains lists instead of depsets
to avoid the number of to_list() calls.

Replaces all ctx.files.pre/post_defconfig_fragments with access to
DefconfigFragmentsListInfo. This is so that we can modify it to include
fragments from base_kernel in the next change.

Bug: 416803611
Change-Id: I4564b6e5e1e5a54b396a8374a5f9d545e298c249