kleaf: Inherit post_defconfig_fragments from base_kernel.
For mixed builds, the device kernel_build() should inherit post_defconfig_fragments
from the base_kernel.

This behavior can be disabled by setting
--noincompatible_inherit_post_defconfig_fragments_from_base_kernel.

This is a no-op for device kernel_build()'s that has base_kernel set to
GKI, because GKI's don't have post_defconfig_fragments at this
point.

post_defconfig_fragments was called defconfig_fragments in
android15-6.6. Note that defconfig_fragments in android15-6.6 did not
have the same inheritance, until this patch is backported.
The (deprecated) defconfig_fragments in android16-6.12 is the same
as post_defconfig_fragments, so it is inherited as well.

The following attributes are also implicit post_defconfig_fragments:

- page_size
- sanitizers
- trim_nonlisted_kmi

Inheriting page_size is okay because the
device kernel_build() should have specified `page_size = 16k` prior to
this change if base_kernel = kernel_aarch64_16k. After this change, any
16k mixed kernel_build()s can delete `page_size = 16k`. But specifying
it explicitly is also okay, because duplicated post_defconfig_fragments
is fine.

Inheriting sanitizers is okay because GKI doesn't have sanitizers set.

As a special case,
trim_nonlisted_kmi is placed in post_defconfig_fragments_non_inherited
list, so it is NOT inherited; its default value is the same as before
this CL.

Bug: 416803611
Change-Id: I6b25c3dd14c324004ba309fecd1080d76457363a