Add blocking IO interfaces to BlockIo
Introduces read_blocks_sync and write_blocks_sync interfaces to the
BlockIo trait that can be overriden by implementation. The purpose is
to allow backend to provide separate and optimized implementation
for blocking IO. This is motivated by the observation that on some
platforms, non-blocking IO introduces additional complexity and
overhead due to handling of interrupt, synchronization, periodic
polling etc, which can not be neglected when reading small amount of
data. Allowing backend to override non-blocking IO mitigates the issue.

Saves about 90~100ms for syncing GPT on sorrel.

Bug: 417987319

Change-Id: Id44548a63ef0b0c3aa4b2ae6afa46d9dea7f78f3