ANDROID: fscrypt: reconcile with v6.16-rc1
"fscrypt: add support for hardware-wrapped keys" is now upstream; see
https://git.kernel.org/linus/14f19dc6440f23f4.  Reconcile the Android
kernel with a minor cleanup made in the final version of the patch.
Also remove a paragraph from the comment in fs/crypto/hkdf.c that is no
longer present upstream and should have been removed in an earlier merge
resolution.

Bug: 160883801
Change-Id: I90247389b57bbf92165a5a1d42352e0f5f361646
Signed-off-by: Eric Biggers <ebiggers@google.com>