kleaf: allow kernel_build outs to be configurable.
This change adds the kernel_build(generate_out_targets=).
Default is true and may be
changed to False in the future. When it is true, labels like
kernel_aarch64/vmlinux etc. are generated; this is the legacy
behavior. This only works when `outs`/`module_outs` etc. is not
a select() expression.

Bug: 419603781
Change-Id: I6f5422309ec95cf3bcc49aba8a0f47daef1b40fc