UPSTREAM: wifi: remove zero-length arrays
All of these are really meant to be variable-length, and
in the case of s1g_beacon it's actually accessed. Make that
one in particular, and a couple of others (that aren't used
as arrays now), actually variable.

Reported-by: syzbot+fd222bb38e916df26fa4@syzkaller.appspotmail.com
Fixes: 1e1f706fc2ce ("wifi: cfg80211/mac80211: correctly parse S1G beacon optional elements")
Link: https://patch.msgid.link/20250614003037.a3e82e882251.I2e8b58e56ff2a9f8b06c66f036578b7c1d4e4685@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

Bug: 516582219
Change-Id: I2e8b58e56ff2a9f8b06c66f036578b7c1d4e4685
(cherry picked from commit a85b8544d46390469b6ca72d6bfd3ecb7be985ff)
Signed-off-by: Yichen Yu <yichenyu@google.com>