kleaf: remove trailing comments before using merge_config.sh.
Some defconfig fragments contains trailing comments, e.g. for # nocheck.
They should not be present in the final .config because they are
recognized as part of a string value.

Impl note: I could also use `perl -pe` for the regex, but perl is not
hermetic yet and hence technically more error-prone than using
the hermetic Python.

Bug: 424212284
Change-Id: I1dbdb16f502d5839c396cf25bb4e091c225f8c33