ANDROID: [TEMP] GKI: x86: Update INT340X modules
The platform temperature control interface driver is added to int340x
thus modules list needs to be updated.

Bug: 393308727
Upstream-Task: 394562775
Link: https://patch.msgid.link/20250429000110.236243-2-srinivas.pandruvada@linux.intel.com

Recent changes on the android-mainline branch conflicted with temporary
patches on the andorid-mainline-desktop branch, thus they are reverted
and reapplied after merge. Eventually they are going to be reverted too.
Related patches for the upstream are currently under review.

Bug: 448988819
Link: https://android-review.googlesource.com/c/kernel/common/+/3783531
Link: https://android-review.googlesource.com/c/kernel/common/+/3666831
Link: https://lore.kernel.org/all/20251002113404.3117429-1-srosek@google.com/
Bypass-Upstream-Divergence-Tracking

Change-Id: If5139f6ea6ab5a95372a3ac39c8036af4b8001e4
Signed-off-by: Slawomir Rosek <srosek@google.com>