ANDROID: rust_binder: allow PollTable to be null
Rust Binder currently makes a nullptr deref when the poll_table passed
to f_ops->poll is a null pointer. This is due to the incorrect
assumption that this pointer is never null. To fix this, I adjusted the
API of the Rust PollTable to allow turning null ptrs into a Rust
PollTable and adjusted Rust Binder to use the updated API. By also
adjusting the PollTable api itself, and not just Rust Binder, this kind
of mistake should be prevented in the future.

Bug: 426545861
Change-Id: I1eabd62d5e499c83b990517c93a63d7de49252ab
Signed-off-by: Alice Ryhl <aliceryhl@google.com>