s2mpg11-regulator: use symbolic name for OCP interrupt
In this driver, buck_ocp_irq[] is allocated using

    s2mpg11->buck_ocp_irq = devm_kzalloc(&pdev->dev,
                                         sizeof(int) * S2MPG11_BUCK_MAX,
                                         GFP_KERNEL);

so the loop in the OCP interrupt handler should use the same macro to
limit access to the array.

Note that S2MPG11_BUCK_MAX resolves to 12, so this commit is not fixing
an out-of-bounds access.

On the other hand, s2mpg11 has actually 13 bucks that can generate OCP
interrupts, but that is out of scope for this change.

Test: TH
Bug: 427668545
Change-Id: I53e05df8652fbc12385c2125b5924ec13575c64b
Signed-off-by: André Draszik <draszik@google.com>