ANDROID: pkvm: vmx: Disable IRQs on first CPU in this_cpu_do_finalise_hc()
Commit 403e5882f5db ("ANDROID: pkvm: vmx: Remove irq disable/enable in
this_cpu_do_finalise_hc()") removed disabling/enabling IRQs on all CPUs,
claiming that this_cpu_do_finalise_hc() already runs in an interrupt
(IPI) context. But that is not true for the first CPU which runs pkvm
init and sends those IPIs to all other CPUs.

The reason for introducing this IRQs disable/enable in commit
8ad35cd28d66 ("ANDROID: pkvm: x86: init: add percpu pkvm_enabled flag")
was to prevent the IOMMU driver's IRQ handler from running on this CPU
when the finalize hypercall is already done (so the CPU is fully
deprivileged and thus cannot access the IOMMU MMIO) but pkvm_enabled
is not set yet (so the IRQ handler wouldn't know that it should use
hypercalls instead of accessing the IOMMU MMIO directly). So, since this
may still happen on the first CPU, restore disabling/enabling IRQs
(but on the first CPU only).

Bug: 391539939
Test: boot
Upstream-Task: 402758258

Fixes: 403e5882f5db ("ANDROID: pkvm: vmx: Remove irq disable/enable in this_cpu_do_finalise_hc()")
Change-Id: Ifbabdc10d6ea43476350da5fbe47551fac42e191
Signed-off-by: Dmytro Maluka <dmaluka@google.com>