ANDROID: pkvm: vmx: Protect most of hyp memory in debug mode
Since commit c50f94481e10 ("pkvm: vmx: Not to protect pkvm text in debug
mode"), when in debug mode, the host has access to the entire memory
reserved for the pKVM hypervisor, in order to allow using dynamic debug,
ftrace etc for the hypervisor code. But for that it only needs access to
the hypervisor's code and data sections, not necessarily to its memory
allocation pools or the pvmfw memory. So give it only access to what it
needs, for reducing the difference between debug and non-debug modes and
for better possibilities for testing pKVM memory protection in debug
mode.

Bug: 391539939
Test: 1. dynamic_debug and ftrace works for pkvm hyp functions. 2. when
reading hyp memory (e.g. at address from dmesg | grep 'kvm.*Reserved')
or pvmfw memory via /dev/mem, garbage is returned and pKVM prints errors
about ept violations.
Upstream-Task: 402758258

Fixes: c50f94481e10 ("pkvm: vmx: Not to protect pkvm text in debug mode")
Change-Id: Ifd3a1a738ccf691d04bb69c9e7e0b046995a1908
Signed-off-by: Dmytro Maluka <dmaluka@google.com>