Rework get buffer API in GblOps
get_image_buffer() currently has an issue that it allocates and returns
completely different buffers each additional time it is called, even
for the same input. Previously returned buffers are silently leaked.
This makes the API highly error-prone, especially when it is used to
obtain vendor provided memory for loading images or preloaded images.

The main cause is that the API transfer buffer ownership to caller and
therefore by design disallows backend to recycle and return the same
buffers. The CL introduces an alternative approach that instead let
implementation manages ownershihp. This is done by returnning an opaque
type defined by backend that can implement Drop. This allows us to
enforce the requirement that backend should always return the same
buffer or errors out if a previous instance has not dropped.

Added support for vendors to pass preloaded partitions using the new
API.

Bug: 430068343

Change-Id: Iad2f1bdd8c16f83cb2d8e02034b8f31337512245