ANDROID: DEBUG: pkvm: x86: Add per-vm pkvm_vmexit_trace
Create a per-vm pkvm_vmexit_trace file in the per-vm sub-folder under
/sys/kernel/debug/kvm/. Cat this file will only show the pkvm trace data
of this vm.

The functions of the files in /sys/kernel/debug/pkvm are remaining the
same, which controls the pkvm trace on/off globally and shows the global
trace as an overview. Emphasized this by renaming the dump_vmexit_trace
to vmexit_trace_overview.

Upstream-Task: 402758258

Fixes: c54a3b1bcd04 ("ANDROID: DEBUG: pkvm: x86: add vmexit trace support")
Change-Id: I615072650a0c111399e7675df5cb45d7ad9076b3
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>