GKI: compare_builds: evaluate artefacts using DFS
The previous algorithm was a hacked-up flood fill. The intent is much
more clearly reflected using a straightforward memoised DFS.

A minor disadvantage of this approach is the explanatory "this node is
good <- that node is good" annotations have been lost. They could be
added back with a little more work.

Change-Id: Ifd1d4a9c21de16619c83130062f675d88ae619a4
Signed-off-by: Giuliano Procida <gprocida@google.com>