ANDROID: sched: Remove set_blocked_on_waking_nested() warning
In e740e8d9eab4 ("ANDROID: Sync proxy-exec logic to v19 (from v18)"),
I consolidated logic in both __ww_mutex_wound() and __ww_mutex_die()
to use a new set_blocked_on_waking_nested() helper.

This helper included a WARN_ON() used in __ww_mutex_die(), but this
assertion isn't generally true in the __ww_mutex_wound() case, but
adding the helper caused it to be used in both.

So drop the warning from the set_blocked_on_waking_nested() helper.

This was already addressed in the v20 series of the proxy-exec patch
set, along with other cleanups, but I missed that I should port it
to andorid16-6.12.

Since the v20 change is a little more complex, and vendors are
hitting this right now, this patch simply drops the incorrect
warning.

Bug: 434841783
Fixes: e740e8d9eab4 ("ANDROID: Sync proxy-exec logic to v19 (from v18)")
Change-Id: I8f611b60b855cdf81af84ad9c575117340431848
Signed-off-by: John Stultz <jstultz@google.com>