kleaf: change algorithm to find ROOT_DIR real path
Files in a sandbox are symlinked to real files in the project. When such
file is used in debug information, they are resolved to a real path
outside of ROOT_DIR, which is leaked to kernel strings and breaks
reproducibility of the kernel. Previous algorithm to find equivalent of
ROOT_DIR for such files was broken.

This change adds another algorithm to find ROOT_DIR real path using some
file from kernel sources, which realpath should lead us to a real path
of ROOT_DIR.

Also sandbox directories usually are subdirectories of
KLEAF_INTERNAL_ROOT_DIR_REALPATH, and if multiple remappings are given
and several of them match, the last matching one is applied. So we need
to have KLEAF_INTERNAL_ROOT_DIR_REALPATH remapping before ROOT_DIR
remapping.

Bug: 389927486
Change-Id: Ib1f57de81ff06ad4d7766e26b89152169b4603cf
Signed-off-by: Aleksei Vetrov <vvvvvv@google.com>