kleaf: Fix integration test with init_ddk call.
A bug was found when rewriting executable_dispatcher. In particular,
if we run the following in this sequence:

bazel build //build/kernel/kleaf/tests/integration_test
bazel build //build/kernel:init_ddk

After the first command,
<output_base>/external/+kernel_toolchain_ext+kleaf_host_tools/rsync is
present because it is part of the dependency. However, after the second
command, external/+kernel_toolchain_ext+kleaf_host_tools/rsync
disappears partly because the repo is not fetched because it is not part
of the dependency.

To mitigate the problem, let integration_test depend on the init_ddk
target, then run the binary from RUNFILES_DIR directly, instead of
calling bazel again.

Change-Id: Ic1e368098d82946889d16d4770e67976acb0ee6c