kleaf: Add devicetree_toolchain.
With this change, a toolchain with just DTC is added. fdtoverlay
is added only when --dtc_from_sources, which is the default.

Flags are forked from Makefile.dtbs. We do not plan to expand this list
or sync this list with Makefile.dtbs to ensure as many warnings are
emitted as possible.

Bug: 376513172
Change-Id: Ifbda9a020632fbd67c45281bdabdf76e65b1f92f