Add documentation for GBL EFI Fastboot local session methods
Device manufacturers may wish to provide a graphical boot menu or
fastboot session that the user can interact with via the screen and
physical buttons. The GblEfiFastbootProtocol enables this
functionality via the `StartLocalSession`, `UpdateLocalSession`, and
`CloseLocalSession` methods.

Add documentation for the methods prototypes and semantics to the
GblEfiFastbootProtocol markdown doc.

Change-Id: I67087bc3c10f92c2ea60b7ec2f88864a06959f58