ANDROID: PM: Support to abort suspend during filesystem sync
This patch adds functionality to handle a wakeup event in the filesystem
sync phase of suspend. The intent is to allow aborts during fs_sync,
which can take 15+ seconds; during this period, from the user's
perspective it looks like the device is unresponsive to any form of
input.

Additionally, there is a new sysfs node added at
/sys/module/suspend/parameters/pm_fs_abort that controls this feature.
By default, this feature is off.

Link: https://lore.kernel.org/all/20250821004237.2712312-1-wusamuel@google.com/
Bug: 315272811
Change-Id: I514627feb9a365062171913b7be3aa433cfc7add
[Added sysfs node to control the feature; resolved conflict in
kernel/power/suspend.c for header includes]
Signed-off-by: Samuel Wu <wusamuel@google.com>