ANDROID: GKI: Disable KABI macros for !CONFIG_64BIT
GKI currently maintains a stable ABI only for the ARM64 architecture,
however the KABI macros take effect on all architectures. This causes
padding to be added for those architectures which is unnecessary, but
not usually harmful. However if the alignment or size checks in
_ANDROID_KABI_NORMAL_SIZE_ALIGN fail, they will cause a build failure on
the basis of an ABI break. That should only occur on architectures where
we maintain a stable ABI. While ARM64 is the only currently supported
architecture, we may support other 64 bit architectures in the future,
so disable the KABI macros for non-64 bit builds.

Bug: 441915167
Change-Id: I6db6991c79140aa084bbf560d9c0f2e3c28d782e
Signed-off-by: T.J. Mercier <tjmercier@google.com>