ANDROID: pkvm: vmx: Improve and fix up NMI injection to host
Backport improvements of NMI handling in pKVM hypervisor, implemented
by Chuanxiao in [1] as a part of the v6.15 rebase of pKVM-IA.

Namely, ensure that NMIs are injected into the host in an
architecturally correct way:

- If there are multiple NMIs in a row, make sure to inject them all, by
  maintaining a counter of queued NMIs.

- NMI should immediately preempt any other exception handling. To achieve
  that, request immediate vmexit after vmentry, using preemption timer
  with timeout set to 0, rather than just rely on the irq window.

Besides the backport, this patch also fixes a regression introduced in
commit fcd206cc6624 ("ANDROID: pkvm: vmx: Inject #GP into host on failed
EPT violation"): when injecting #GP into the host due to the host's
attempt to access protected memory, if there is also a pending NMI to be
injected, the #GP overwrites the NMI in VM_ENTRY_INTR_INFO_FIELD, so the
NMI is not injected in this case.

[1] https://github.com/intel-staging/pKVM-IA/commit/7bf2c58d0de943ab7d047e8e039d00e666a61a13

Bug: 349990461
Upstream-Task: 402758258
Test: 1. run pKVM-IA in qemu, use "nmi" command in qemu monitor to
inject NMIs into the qemu VM, and observe its dmesg.
2. hack pkvm_vmexit_main(): add "for (i = 0; i < 1000000; i++)
cpu_relax();" to make vmexit handling much longer and thus increase the
likelihood of NMI triggering in pKVM, not in the host.
3. connect to the qemu VM with gdb, set breakpoint on handle_nmi(), and
inject NMIs until the breakpoint is triggered (i.e. NMI occurred in
pKVM); inject another NMI while handle_nmi() is interrupted by gdb, and
run "continue" in gdb; repeat several times; run final "continue" in gdb
and observe in the qemu VM's dmesg that the host receives all pending
NMIs.

Fixes: fcd206cc6624 ("ANDROID: pkvm: vmx: Inject #GP into host on failed EPT violation")
Change-Id: Ib4bab35797e32a90584f61eda024c6df5b8d657b
Signed-off-by: Dmytro Maluka <dmaluka@google.com>