kleaf: Add INFO logs to --config=silent.
Warnings printed to stderr inside actions are reflected as INFO logs
at the Bazel level. Hence, we have to add INFO logs to
--ui_event_filters for these warnings to be visible even under
--config=silent, ensuring that we have no warnings from our actions.

However, adding --ui_event_filters=info adds a bunch of other INFO logs
from Bazel. So we have to add them to the regex allowlist.

Also update a comment in the regex allowlist.

Also add a few known violations that are yet to be investigated.

Bug: 445484229
Change-Id: I5887759d86877f5caf5273ed80cdc725dacc6871