FROMLIST: selftests/net: psock_tpacket: Fix null argument warning in walk_tx
The sendto() call in walk_tx() was passing NULL as the buffer argument,
which can trigger a -Wnonnull warning with some compilers.

Although the size is 0 and no data is actually sent, passing a null
pointer is technically incorrect.

This commit changes NULL to an empty string literal ("") to satisfy the
non-null argument requirement and fix the compiler warning.

Bug: 444295917
Change-Id: If8deb520e65713c1a751ee4aa30ca40cc91168d9
Link: https://lore.kernel.org/all/20250924044142.540162-1-wakel@google.com/
Signed-off-by: Wake Liu <wakel@google.com>
Test: bazel build //common:kernel_aarch64_dist