ANDROID: arm64/dts: exynos: gs101: add androidboot.boot_device
This property is needed to tell Android the name of the UFS boot device.
This needs to be in the device tree because the upstream device tree
has a slightly different path to the boot device than the downstream
device tree.

Keeping this property in the device tree allows us to use the same
Android userspace images for both the upstream and downstream kernel
images.

The downstream sysfs path is:

  /sys/devices/platform/14700000.ufs

while the upstream sysfs path is:

  /sys/devices/platform/soc@0/14700000.ufs

Bug: 445758592
Change-Id: I5afa4d87e8fc2f563302f16148f66b931123cd64
Signed-off-by: Will McVicker <willmcvicker@google.com>