ANDROID: simplify kgdb targets.
kernel_build.module_outs now accepts a select() expression, with the
cost that <kernel_build_name>/<out> targets (e.g. rockpi4/vmlimux)
are no longer generated. We do not rely on these targets for
db845c, rockpi4 and yukawa, so we can safely enable the use
of select() in module_outs, which greatly simplifies the enabling of
watchdog conditioned on --kgdb.

Bug: 419603781
Change-Id: I3e9bb0e2ac6e89522075384f2bcb304b3d9f642e
Signed-off-by: HONG Yifan <elsk@google.com>