UPSTREAM: platform/chrome: cros_usbpd_notify: defer probe when parent EC driver isn't ready
The cros-usbpd-notify-acpi probe currently does not exit when it fails
to get a pointer to the ChromeOS EC device. It is expected behavior on
older devices, where GOOG0004 is not a parent of GOOG0003.

Update the cros-usbpd-notify-acpi probe to check for a GOOG0004 parent
fwnode. If the device has correct device hierarchy and fails to get an
EC device pointer, defer the probe function.

Signed-off-by: Jameson Thies <jthies@google.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Link: https://lore.kernel.org/r/20251007004043.4109957-1-jthies@google.com
Signed-off-by: Tzung-Bi Shih <tzungbi@kernel.org>
(cherry picked from commit e4ee0bb077cd7d70207647a0106f6ea6a74c2636)

Bug: 437116596
Test: cros_ec_ucsi probe succeeds

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: Ifa92676897d0f9c9a53d1f88101f4156389f51b2