FROMLIST: rust_binder: report freeze notification only when fully frozen
Binder only sends out freeze notifications when ioctl_freeze() completes
and the process has become fully frozen. However, if a freeze
notification is registered during the freeze operation, then it
registers an initial state of 'frozen'. This is a problem because if
the freeze operation fails, then the listener is not told about that
state change, leading to lost updates.

Bug: 448820695
Link: https://lore.kernel.org/r/20251007-binder-freeze-v2-3-5376bd64fb59@google.com
Change-Id: I68bcca4d63df6e89bf958a90d4ae353a8ea0b501
Signed-off-by: Alice Ryhl <aliceryhl@google.com>