ANDROID: GKI: x86: Disable vsyscall emulation
This enables emulation of the legacy vsyscall page.  Disabling
it is roughly equivalent to booting with vsyscall=none, except
that it will also disable the helpful warning if a program
tries to use a vsyscall.  With this option set to N, offending
programs will just segfault, citing addresses of the form
0xffffffffff600?00.

Disabling this option saves about 7K of kernel size and
possibly 4K of additional runtime pagetable memory.

Bug: 384758871

Change-Id: Id0669e4f7984282322ee4e4c12f486a906342803
Signed-off-by: Jeff Xu <jeffxu@google.com>
Signed-off-by: Slawomir Rosek <srosek@google.com>