Revert "ANDROID: PM / Domains: add vendor_hook to disallow domain idle state"
This reverts commit e1148714015118450923dd896a46d1a21a33b2d7.

The hook android_vh_allow_domain_state is not used by any vendor, so
remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 175718935
Bug: 195914333
Cc: Lina Iyer <ilina@codeaurora.org>
Cc: Maulik Shah <mkshah@codeaurora.org>
Change-Id: I18d6f3c5fac5f8709aab28d6f1fe0abc40cfdfff
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>