Revert "ANDROID: GKI: net: add vendor hooks net qos for gki purpose"
This reverts commit 6c1c18fcb8b7f3fd7f26ab075b96e9a135d2faa3.

The hook android_rvh_tcp_rcv_spurious_retrans is not used by any vendor,
so remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 351960112
Bug: 356582738
Bug: 356617512
Bug: 422600631
Cc: Jyu Jiang <jyu.jiang@vivo.corp-partner.google.com>
Change-Id: I35d3c2fd7509351ec95ab9f49004f28a9880d03d
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>