Revert "ANDROID: gic: Add affinity-setting vendor hooks to gic"
This reverts commit 3dc71aaa72d7daabe5c17670e31b4034c4fed18b.

The hook android_vh_gic_set_affinity is not used by any vendor, so
remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 266450121
Cc: Guru Das Srinagesh <quic_gurus@quicinc.com>
Change-Id: I787d5d0b1fee60d4cb275efc1be83b992edf4880
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>