Revert "ANDROID: add vendor hooks for getting information of dtask"
This reverts commit e1456ad867a392906405d934b0e27c9fde3119f1.

The hooks:
	android_vh_dpm_wait_start
	android_vh_dpm_wait_finish
	android_vh_sync_irq_wait_start
	android_vh_sync_irq_wait_finish
	android_vh_flush_wq_wait_start
	android_vh_flush_wq_wait_finish
	android_vh_flush_work_wait_start
	android_vh_flush_work_wait_finish
are not used by any vendor, so remove them to help with merge issues
with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 372142930
Bug: 386837579
Cc: Sangmoon Kim <sangmoon.kim@samsung.com>
Change-Id: Ica1f5433a5ed7980edcf3342a2d6b04b95f2feb1
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>