avb: rectify avb_crypto_ops_impl dependency.
The includes=../gbl/libavb is very fragile, because once gbl changes from
a repository to a bazel module, the path changes from external/gbl to
external/gbl+. The pure name 'gbl' is not bzlmod ready.

To properly fix this, define a cc_library that contains the header
with the proper includes.

Bug: 413506174
Change-Id: I9454ad94d97f14e5bdae6e063393925dc44f586d