FROMLIST: kunit: prevent log overwrite in param_tests
When running parameterized tests, each test case is initialized with
kunit_init_test(). This function takes the test_case->log as a parameter
but it clears it via string_stream_clear() on each iteration.

This results in only the log from the last parameter being preserved in
the test_case->log and the results from the previous parameters are lost
from the debugfs entry.

Fix this by manually setting the param_test.log to the test_case->log
after it has been initialized. This prevents kunit_init_test() from
clearing the log on each iteration.

Fixes: 4b59300ba4d2 ("kunit: Add parent kunit for parameterized test context")
Signed-off-by: Carlos Llamas <cmllamas@google.com>

Bug: 454945983
Link: https://lore.kernel.org/all/20251024190101.2091549-1-cmllamas@google.com/
Change-Id: I34bf479d787777f6837b8d0f4885a3d94b7ccb33
Signed-off-by: Carlos Llamas <cmllamas@google.com>