ANDROID: KVM: arm64: Fix pKVM ftrace va to phys
Ftrace needs to resolve module VA's to phys which can be only done by
walking the hyp page-table (modules are mapped into the private range).
Those VA's are solely coming from the .hyp.text ELF sections so we can
easily differentiate if the VA is from a module or not by looking at
__hyp_text_start.

However, we are not guaranteed that .hyp.text will be placed after the
private range. Fix pKVM ftrace module patching by also checking the
.hyp.text section end boundary.

Bug: 422057361
Bug: 357781595
Change-Id: If468f432b737ebf2317682db3f25819e58193c9a
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>