ANDROID: module: allow MODULE_SIG_PROTECT_LIST to be relative against $(objtree)
Unlike CONFIG_UNUSED_KSYMS_WHITELIST in
commit a2e3c811938b ("module: allow UNUSED_KSYMS_WHITELIST to be relative against objtree.")
which are only used for building the kernel and in-tree modules,
MODULE_SIG_PROTECT_LIST is also used for building external
modules (KBUILD_EXTMOD is set).

When building external modules,
$(objtree) is no longer `.`, so the
$(wildcard $(MODULE_SIG_PROTECT_LIST)) no longer properly detects
it relative to $(objtree).

Hence, we cannot mechanically apply
commit a2e3c811938b ("module: allow UNUSED_KSYMS_WHITELIST to be relative against objtree.")
to MODULE_SIG_PROTECT_LIST, but instead explicitly use $(objtree)
in the wildcard.

Bug: 448822709
Bug: 333769605
Change-Id: I8d42ef759d6990cf69e665ad24bda89d957cfc4a
Signed-off-by: HONG Yifan <elsk@google.com>