ANDROID: Update AutoFDO profile for 6.12.52
The new profile is collected from Pixel 10 instead of Pixel 6. It
aggregated more samples (8e12 vs 6e12) and found more hot functions
(3002 vs 2610). The profile is made smaller by removing samples for most
cold functions, which only takes 0.4% of samples.

For the size of .text in vmlinux:
no profile is 18.9M, with old profile is 19.2M, with new profile is
19.4M.

For testing, we added Pixel 10 and Pixel 8, instead of solely relying on Pixel 6.

Bug: 322399357
Test: Benchmarks run on Pixel 6, 8, and 10 show performance improvements:
  - Boot time: 2.1%
  - Cold App launch time: 4.3%
  - Binder-rpc: 21.7%
  - Binder-addints: 12.3%
  - Hwbinder: 20.0%
  - Bionic (syscall_mmap): 9.3%

Change-Id: I112b236bbb1660ae5eb24b6125a82b9db485797f
Signed-off-by: Yabin Cui <yabinc@google.com>