ANDROID: Update AutoFDO profile for 6.6.111
The new profile is collected from Pixel 10 instead of Pixel 6. It
aggregated more samples (8e12 vs 6e12) and found more hot functions
(2962 vs 2482). The profile is made smaller by removing samples for most
cold functions, which only takes 0.1% of samples.

For the size of .text in vmlinux:
no profile is 16.5M, with old profile is 16.9M, with new profile is
17.1M.

For testing, we added Pixel 10, instead of solely relying on Pixel 6.

Bug: 322399357
Test: Benchmarks run on Pixel 6 and 10 show performance improvements:
  - Boot time: 2.1%
  - Cold App launch time: 3.7%
  - Binder-rpc: 6.7%
  - Binder-addints: 14.5%
  - Hwbinder: 7.8%
  - Bionic (syscall_mmap): 3.8%

Change-Id: I744372cdbe7edf5104d6c278d854347fc5731311
Signed-off-by: Yabin Cui <yabinc@google.com>