Merge 305b1a39f3bb ("sched/deadline: Fix race in push_dl_task()") into android16-6.12-lts
Steps on the way to 6.12.54

Merges away this commit due to all of the Android-specific changes that
have been done in the deadline scheduler, and this commit does not apply
any more.  If it needs to come back in the future, it can be done so in
an Android-specific way.

Resolves merge conflicts in:
	kernel/sched/deadline.c

Change-Id: I0ac04e3e80837c86f2cb5db8a9592f51b038bec1
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>