ANDROID: KABI: fix up abi change in struct usb_request
In commit 97fc7aa654e3 ("usb: gadget: Store endpoint pointer in
usb_request"), a new pointer is added to struct usb_request to fix some
issues with usb gadget drivers.  As we want those fixes in the
android14-6.1 tree, use a reserved kabi field for this new pointer to
preserve the abi.

Fixes: 97fc7aa654e3 ("usb: gadget: Store endpoint pointer in usb_request")
Change-Id: Ieacf1186052a20c43442576e6d8695e86a2189ed
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>