ANDROID: selinux: fix insufficient buffer for memfd_class name
When constructing the SELinux capabilities information string, the
"memfd_class" string was not taken into consideration for buffer size
calculation. As memfd_class is a capability Android backported from
upstream, we neglected to replicate the necessary buffer size
calculation logic. This bug did not cause an overflow because string
construction was done by functions aware of the available buffer size;
instead, it only resulted in the output string being truncated. This
change fixes this bug.

Bug: 455849277
Test: phone Cuttlefish booted. Memfd capability is enabled.
Fixes: d4564b29b1f9 ("BACKPORT: FROMGIT: memfd,selinux: call security_inode_init_security_anon()")
Change-Id: I9aafe0e79813ddcf88abd95d97b982dcdbf77a44
Signed-off-by: Takaya Saeki <takayas@google.com>