Revert "ANDROID: sched: Add vendor hooks for skipping sugov update"
This reverts commit 8bdea28a375adf11d539582615e3429026332dc1.

The hook android_rvh_set_sugov_update is not used by any vendor, so
remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 170511085
Cc: Wei Wang <wvw@google.com>
Cc: Will McVicker <willmcvicker@google.com>
Change-Id: Ia823501433a21b6e59b9465cce5d30ac62e1af0e
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>