Revert "ANDROID: mm: create vendor hooks for mm  proactive compact"
This reverts commit 891189b643600178d4736622b2ae2014bc951ea4.

The hook android_vh_proactive_compact_wmark_high is not used by any
vendor, so remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 391491611
Cc: yipeng xiang <yipengxiang@honor.corp-partner.google.com>
Change-Id: Ia125c602e91e9fb8bb38e52560e4838b4d7de73d
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>