ANDROID: soc: mediatek: make MTK_INFRACFG tristate
MTK took the effort to modularize for mtk-infracfg
driver for MTK desktop devices and made few other
vendor modules relying on this driver being a module.
It is also critical for these devices to boot.

The MTK_INFRACFG is generally disabled in gki_defconfig
which implies it is not used anywhere. Also all options
that select MTK_INFRACFG are also tristate
and they are set as vendor modules for this specific case.

Bug: 468281238
Bug: 517435391

Change-Id: Iddef1da38c6db22b1e7ea74fc723b6c804184339
Signed-off-by: Zyta Szpak <zyta@google.com>
(cherry picked from commit 3aa7b692360f836deab9fd2a56ecbd3ebe738159)