ANDROID: ashmem: adapt to new mutex pinning rules
Upstream has introduced stricter pinning rules for lock::Guard, and now
DerefMut is no longer implemented for MutexGuard<T> when T is a pinned
type. This leads to build errors with AshmemInner's inner lock usage:

  error: variable does not need to be mutable
     --> drivers/staging/android/ashmem_rust.rs:314:13
      |
  314 |         let mut asma = self.inner.lock();
      |             ----^^^^
      |             |
      |             help: remove this `mut`
      |
      = note: `-D unused-mut` implied by `-D warnings`
      = help: to override `-D warnings` add `#[allow(unused_mut)]`

In order to fix this, add the #[pin_data] attribute to AshmemInner and
declare safe mutable access to the fields of this pinned struct through
the mutex guard.

Cc: Isaac J. Manjarres <isaacmanjarres@google.com>
Fixes: da123f0ee40f ("rust: lock: guard: Add T: Unpin bound to DerefMut")
Change-Id: Ic807daba552891650d11009624733e8e2e433d62
Suggested-by: Alice Ryhl <aliceryhl@google.com>
Signed-off-by: Carlos Llamas <cmllamas@google.com>