Remove match against uninitialized pointer
`of_property_match_string` expects the third parameter to be a string to
search for. We're currently passing in an uninitialized pointer, which
will result in it performing potentially wild reads.

It's not currently hit because the dfs_id is set, so remove the search
entirely just to be safe.

Bug: 457798339
Signed-off-by: Matthew Maurer <mmaurer@google.com>
Cherrypick-From: https://android-review.googlesource.com/q/commit:384242e737e2dda53937b1492e99c144b680f897
Merged-In: I06d02dc844fcaf007081929f8625188c79393d83
Change-Id: I06d02dc844fcaf007081929f8625188c79393d83